renamed acknowledge service in ccd definition and e2e tests
