+++ ccd-definition/AuthorisationCaseEvent/caseAccessAssignment.json
@@ -9,7 +9,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseEventID": "ACKNOWLEDGE_SERVICE",
+    "CaseEventID": "ACKNOWLEDGE_CLAIM",
     "UserRole": "caseworker-caa",
     "CRUD": "CRU"
   },
