+++ ccd-definition/CaseEvent/User/UserEvents.json
@@ -60,9 +60,9 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "ID": "ACKNOWLEDGE_SERVICE",
-    "Name": "Acknowledge service",
-    "Description": "Defendant solicitor is acknowledging service",
+    "ID": "ACKNOWLEDGE_CLAIM",
+    "Name": "Acknowledge claim",
+    "Description": "Defendant solicitor is acknowledging claim",
     "DisplayOrder": 4,
     "PreConditionState(s)": "CREATED",
     "PostConditionState": "*",
@@ -72,7 +72,7 @@
     "CallBackURLSubmittedEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/cases/callbacks/submitted",
     "ShowSummary": "Y",
     "ShowEventNotes": "N",
-    "EndButtonLabel": "Acknowledge service",
+    "EndButtonLabel": "Acknowledge claim",
     "RetriesTimeoutAboutToStartEvent": 0,
     "RetriesTimeoutURLAboutToSubmitEvent": 0,
     "RetriesTimeoutURLSubmittedEvent": 0
