+++ ccd-definition/CaseEventToFields/AcknowledgeClaim.json
@@ -2,7 +2,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseEventID": "ACKNOWLEDGE_SERVICE",
+    "CaseEventID": "ACKNOWLEDGE_CLAIM",
     "CaseFieldID": "respondentDetails",
     "PageFieldDisplayOrder": 1,
     "DisplayContext": "READONLY",
@@ -14,7 +14,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseEventID": "ACKNOWLEDGE_SERVICE",
+    "CaseEventID": "ACKNOWLEDGE_CLAIM",
     "CaseFieldID": "solicitorReferences",
     "CaseEventFieldLabel": "Confirm your details",
     "PageFieldDisplayOrder": 1,
@@ -29,7 +29,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseEventID": "ACKNOWLEDGE_SERVICE",
+    "CaseEventID": "ACKNOWLEDGE_CLAIM",
     "CaseFieldID": "respondent1",
     "CaseEventFieldLabel": " ",
     "PageFieldDisplayOrder": 2,
@@ -42,7 +42,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseEventID": "ACKNOWLEDGE_SERVICE",
+    "CaseEventID": "ACKNOWLEDGE_CLAIM",
     "CaseFieldID": "respondent1ClaimResponseIntentionType",
     "CaseEventFieldLabel": "The defendant intends to:",
     "PageFieldDisplayOrder": 1,
