+++ e2e/pages/acknowledgeClaim/responseIntention.page.js
